Migrating logic out of worker
