+++ src/Rn.DnsUpdater.Core/DnsUpdateRunner.cs
@@ -1,0 +1,96 @@
+using Microsoft.Extensions.DependencyInjection;
+using Rn.DnsUpdater.Core.Config;
+using Rn.DnsUpdater.Core.Enums;
+using Rn.DnsUpdater.Core.Services.Interfaces;
+using Rn.NetCore.Common.Logging;
+using Rn.NetCore.Metrics;
+using Rn.NetCore.Metrics.Builders;
+
+namespace Rn.DnsUpdater.Core;
+
+public interface IDnsUpdateRunner
+{
+  Task ExecuteAsync(CancellationToken stoppingToken);
+}
+
+public class DnsUpdateRunner : IDnsUpdateRunner
+{
+  private readonly ILoggerAdapter<DnsUpdateRunner> _logger;
+  private readonly IHostIpAddressService _addressService;
+  private readonly IConfigService _configService;
+  private readonly IDnsUpdaterService _dnsUpdater;
+  private readonly IHeartbeatService _heartbeatService;
+  private readonly IMetricService _metrics;
+
+  public DnsUpdateRunner(IServiceProvider serviceProvider)
+  {
+    _logger = serviceProvider.GetRequiredService<ILoggerAdapter<DnsUpdateRunner>>();
+    _addressService = serviceProvider.GetRequiredService<IHostIpAddressService>();
+    _configService = serviceProvider.GetRequiredService<IConfigService>();
+    _dnsUpdater = serviceProvider.GetRequiredService<IDnsUpdaterService>();
+    _heartbeatService = serviceProvider.GetRequiredService<IHeartbeatService>();
+    _metrics = serviceProvider.GetRequiredService<IMetricService>();
+  }
+
+
+  public async Task ExecuteAsync(CancellationToken stoppingToken)
+  {
+    // TODO: [DnsUpdateRunner.ExecuteAsync] (TESTS) Add tests
+    while (!stoppingToken.IsCancellationRequested)
+    {
+      await _heartbeatService.TickAsync();
+      var hostAddressChanged = await _addressService.HostAddressChangedAsync(stoppingToken);
+
+      // Decide if we need to update all entries, or just a smaller subset
+      var dnsEntries = hostAddressChanged
+        ? _configService.GetEnabledEntries()
+        : _configService.GetEntriesNeedingUpdate();
+
+      // Update any DnsEntries returned above
+      await UpdateDnsEntries(dnsEntries, stoppingToken);
+
+      // Wait for the next loop
+      await Task.Delay(_configService.CoreConfig.TickInterval, stoppingToken);
+    }
+  }
+
+
+  private async Task UpdateDnsEntries(IReadOnlyCollection<DnsUpdaterEntry> dnsEntries, CancellationToken stoppingToken)
+  {
+    // TODO: [TESTS] (DnsUpdaterWorker.UpdateDnsEntries) Add tests
+    // Ensure that we have something to work with
+    if (dnsEntries.Count == 0)
+      return;
+
+    _logger.LogInformation(dnsEntries.Count == 1
+      ? "Updating 1 DNS entry"
+      : $"Updating {dnsEntries.Count} DNS entries");
+
+    var builder = new ServiceMetricBuilder(nameof(DnsUpdateRunner), nameof(UpdateDnsEntries))
+      .WithCategory(MetricCategory.DnsUpdater, MetricSubCategory.UpdateEntries)
+      .WithCustomInt1(dnsEntries.Count);
+
+    try
+    {
+      using (builder.WithTiming())
+      {
+        foreach (var dnsEntry in dnsEntries)
+        {
+          builder.IncrementQueryCount();
+          await _dnsUpdater.UpdateEntryAsync(dnsEntry, stoppingToken);
+        }
+
+        _configService.SaveConfigState();
+      }
+    }
+    catch (Exception ex)
+    {
+      builder.WithException(ex);
+      _logger.LogUnexpectedException(ex);
+    }
+    finally
+    {
+      await _metrics.SubmitBuilderAsync(builder);
+    }
+  }
+}
